refactor: ProductService 리팩토링
